pyproject.toml
[project]
name = "nba-recap"
version = "0.1.0"
description = "AI-powered NBA game recaps"
requires-python = ">=3.12"
dependencies = ["nba_api>=1.11.3", "anthropic>=0.76.0", "python-dotenv>=1.0.0"]

[project.scripts]
recap = "nba_recap.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]